+++ volume-cartographer/apps/src/vc_fiberlets.cpp
@@ -602,7 +602,7 @@ void printTubeExtractionProfile(
         return encoded.str();
     };
     output << std::setprecision(17)
-           << "fiberlet_extraction_profile version=17"
+           << "fiberlet_extraction_profile version=20"
            << " anchor_elapsed_seconds=" << extraction.anchors.elapsedSeconds
            << " anchor_cpu_seconds=" << anchor.elapsedCpuSeconds
            << " anchor_profiled_seconds=" << anchorProfiledSeconds
@@ -612,15 +612,23 @@ void printTubeExtractionProfile(
            << " anchor_context_cells=" << anchor.contextCells
            << " anchor_work_cells=" << anchor.workCells
            << " anchor_tiles=" << anchor.tiles
-           << " anchor_sampling_groups=" << anchor.samplingGroups
+           << " anchor_sampling_partitions=" << anchor.samplingPartitions
            << " anchor_workers=" << anchor.workers
            << " anchor_sampler_calls=" << anchor.predictionSamplerCalls
+           << " anchor_shared_sampling_batches="
+           << anchor.sharedSamplingBatches
+           << " anchor_max_sampling_batch_voxels="
+           << anchor.maximumSamplingBatchVoxels
            << " anchor_submitted_prediction_voxels="
            << anchor.submittedPredictionVoxels
            << " anchor_unique_tile_prediction_voxels="
            << anchor.uniqueTilePredictionVoxels
            << " anchor_reused_prediction_voxels="
            << anchor.reusedPredictionVoxels
+           << " anchor_max_shared_sample_bytes="
+           << anchor.maximumSharedSampleBytes
+           << " anchor_max_accounted_live_bytes="
+           << anchor.maximumAccountedLiveBytes
            << " anchor_candidate_observations=" << anchor.candidateObservations
            << " anchor_retained_observations=" << anchor.retainedObservations
            << " anchor_support_stencil_cells=" << anchor.supportStencilCells
@@ -637,15 +645,39 @@ void printTubeExtractionProfile(
            << " anchor_cell_processing_seconds=" << anchor.cellProcessingSeconds
            << " anchor_cell_processing_cpu_seconds="
            << anchor.cellProcessingCpuSeconds
+           << " anchor_shared_sampling_seconds="
+           << anchor.sharedSamplingSeconds
+           << " anchor_shared_sampling_cpu_seconds="
+           << anchor.sharedSamplingCpuSeconds
            << " anchor_coordinate_construction_work_seconds="
            << anchor.coordinateConstructionWorkSeconds
            << " anchor_prediction_sampling_work_seconds="
            << anchor.predictionSamplingWorkSeconds
+           << " anchor_tile_sample_copy_work_seconds="
+           << anchor.tileSampleCopyWorkSeconds
            << " anchor_gradient_construction_work_seconds="
            << anchor.gradientConstructionWorkSeconds
            << " anchor_observation_construction_work_seconds="
            << anchor.observationConstructionWorkSeconds
            << " anchor_fitting_work_seconds=" << anchor.fittingWorkSeconds
+           << " anchor_partition_p50_seconds="
+           << anchor.partitionP50Seconds
+           << " anchor_partition_p95_seconds="
+           << anchor.partitionP95Seconds
+           << " anchor_partition_max_seconds="
+           << anchor.partitionMaximumSeconds
+           << " anchor_tile_preparation_p50_seconds="
+           << anchor.tilePreparationP50Seconds
+           << " anchor_tile_preparation_p95_seconds="
+           << anchor.tilePreparationP95Seconds
+           << " anchor_tile_preparation_max_seconds="
+           << anchor.tilePreparationMaximumSeconds
+           << " anchor_cell_processing_p50_seconds="
+           << anchor.cellProcessingP50Seconds
+           << " anchor_cell_processing_p95_seconds="
+           << anchor.cellProcessingP95Seconds
+           << " anchor_cell_processing_max_seconds="
+           << anchor.cellProcessingMaximumSeconds
            << " anchor_fit_invocations=" << fit.invocations
            << " anchor_fit_nonempty_cells=" << fit.nonemptyCells
            << " anchor_fit_weighted_observations=" << fit.weightedObservations
@@ -775,6 +807,16 @@ void printTubeExtractionProfile(
            << " fiberlet_retained_search_nodes=" << paths.retainedSearchNodes
            << " fiberlet_corner_insertion_attempts="
            << paths.interpolationCornerInsertions
+           << " fiberlet_corner_worker_unique_voxels="
+           << paths.cornerWorkerUniqueVoxels
+           << " fiberlet_corner_worker_pages=" << paths.cornerWorkerPages
+           << " fiberlet_corner_page_directory_probes="
+           << paths.cornerPageDirectoryProbes
+           << " fiberlet_corner_same_page_hits="
+           << paths.cornerSamePageHits
+           << " fiberlet_corner_cached_page_hits="
+           << paths.cornerCachedPageHits
+           << " fiberlet_corner_merged_pages=" << paths.cornerMergedPages
            << " fiberlet_unique_sampled_voxels=" << paths.sampledVoxels
            << " fiberlet_interpolated_scoring_points="
            << paths.interpolatedScoringPoints